Extract health controller.
