+++ ts/backend/test/integration/health.controller.spec.ts
@@ -1,8 +1,8 @@
-import { DebugController } from '../../src/app/features/controller/debug.controller.mjs'
+import { HealthController } from '../../src/app/features/controller/health.controller.mjs'
 import { ControllerTestAdapter, setupControllerTestEnvironment } from '../controller.test-adapter.mjs'
 import { getTestConfig } from './setup.mjs'
 
-describe('Debug Controller Failing controller', () => {
+describe('Health Controller Failing controller', () => {
   let controller: ControllerTestAdapter
 
   test('should because of unreachable broker', async () => {
@@ -10,7 +10,7 @@ describe('Debug Controller Failing controller', () => {
     testConfig.amqp.host = "nix.amqp.com"
     testConfig.amqp.port = 2
     setupControllerTestEnvironment(testConfig)
-    controller = new ControllerTestAdapter(DebugController, testConfig)
+    controller = new ControllerTestAdapter(HealthController, testConfig)
     const res = await controller.testGet('/health/live/', {})
     console.log('= /health/live =')
     console.log(res.status)
@@ -23,7 +23,7 @@ describe('Debug Controller Failing controller', () => {
     testConfig.postgres.host = "nix.postgres.com"
     testConfig.postgres.port = 1 // do not use 0 - would fall back to default
     setupControllerTestEnvironment(testConfig)
-    controller = new ControllerTestAdapter(DebugController, testConfig)
+    controller = new ControllerTestAdapter(HealthController, testConfig)
     const res = await controller.testGet('/health/live/', {})
     console.log('= /health/live =')
     console.log(res.status)
@@ -34,7 +34,7 @@ describe('Debug Controller Failing controller', () => {
   test('should return healthy', async () => {
     const testConfig = await getTestConfig()
     setupControllerTestEnvironment(testConfig)
-    controller = new ControllerTestAdapter(DebugController, testConfig)
+    controller = new ControllerTestAdapter(HealthController, testConfig)
     const res = await controller.testGet('/health/live/', {})
     console.log('= /health/live =')
     console.log(res.status)
